sched: walt: Correct WALT window size initialization
It is preferable that WALT window rollover occurs just before a tick,
since the tick is an opportune moment to record a complete window's
statistics, as well as report those stats to the cpu frequency governor.
When CONFIG_HZ results in a TICK_NSEC that isn't a integral number, this
requirement may be violated. Account for this by reducing the WALT window
size to the nearest multiple of TICK_NSEC.

Commit 578db5dbecfa ("ANDROID: sched: walt: fix window misalignment when
HZ=300") attempted to do this but it is hardcoded specifically for a 20
ms WALT window and CONFIG_HZ=300 and will not work for any other
scenario.

Also, change the type of 'walt_disabled' to bool.

Change-Id: Ie3dcfc21a3df4408254ca1165a355bbe391ed5c7
Signed-off-by: Vikram Mulukutla <markivx@codeaurora.org>
Signed-off-by: kdrag0n <dragon@khronodragon.com>
Signed-off-by: Chenyang Zhong <zhongcy95@gmail.com>
Signed-off-by: ALKALiKong <alikongalk@gmail.com>